in refactoring, basis working
